+++ packages/orq-rc/RELEASES.md
@@ -6439,3 +6439,13 @@ Based on:
 - [typescript v4.3.0-rc.28] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.28] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.28 - packages/orq-rc
+
+## 2026-02-05 13:13:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.705.0 (2.808.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.29] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.29 - packages/orq-rc